Add ButtonID and DisplayPanelID to enumerate each properly
DisplayPanelID will let us differentiate between which panel we are addressing without having to do a String match (let the compiler check that we used a valid ID instead).
ButtonID will do this too, but will also provide a package for the asset name, its hover text, and its command so the DisplayPanels only have to pick the right ButtonID, to get those.