All Menu functions work
Menu builds a new game properly as the user selects the radio buttons and changes the tile colors, and then passes that new game on to the TileGame controller which makes the GameController create a new game with those settings, and then switches the view to the game.

Clicking on Back, goes back to the game.
Clicking on Quit, quits the game.