Major refactor/redesign to remove foolish consistency
Was trying to make everything send "Messages" to each other so there would be few public interface methods, but trying to shoehorn the data into the messages as objects and then cast them to the appropriate type was awkward at best, and the straw that broke the camel's back was trying to get the color list to InfoDisplay - need to provide 2 pieces of information but the framework only allowed one.  If we added one more generic object field (or even the specific data), 99% of the program would ignore it and pass "null", which was already too prevalent in the code.
So now the class hierarchies will still support some basic common functionality (controllers will react(), hover() and unhover() when buttons are interacted with, for example), but the controllers and display panels will interact with each other in a more personalized manner.